+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.23.14
+
+- Fix: use `lexeme` in `package:analyzer` to properly detect conditional imports.
+
 ## 0.23.13
 
 - Detect legacy Kotlin Gradle Plugin application and `kotlinOptions` in Android plugins.